requirements.txt
python-dotenv==1.2.1
ddgs==9.8.0
duckduckgo-search==8.1.1
langchain==0.3.28
langchain-community==0.3.31
langchain-groq==0.3.8
langgraph==0.6.11
beautifulsoup4==4.14.3
requests==2.32.5
deepeval==3.9.7